pyproject.toml
[tool.poetry]
name = "overdose"
version = "0.1.0"
description = "data journalism framework"
authors = ["Emma Brown <nworb95@gmail.com>"]
license = "MIT"
packages = [
    { from = "src", include = "app"},
    { from = "src", include = "bokeh_plotter"},
    { from = "src", include = "database_handler/cornell"},
    { from = "src", include = "database_handler/socrata"},
    { from = "src", include = "database_handler/utils"}
]

[tool.poetry.dependencies]
python = ">=3.6.1,<4"
appdirs = "^1.4.3"
attrs = "^19.3.0"
certifi = "^2019.11.28"
chardet = "^3.0.4"
Click = "^7.0"
elasticsearch = "^7.5.1"
idna = "^2.9"
numpy = "1.18.1"
pandas = "^1.0.1"
pathspec = "^0.7.0"
psycopg2 = "^2.8.4"
python-dateutil = "^2.8.1"
pytz = "^2019.3"
requests = "^2.23.0"
scipy = "^1.5.2"
six = "^1.14.0"
sodapy = "^2.0.0"
SQLAlchemy = "^1.3.13"
toml = "^0.10.0"
typed-ast = "^1.4.1"
urllib3 = "^1.25.8"
xlrd = "^1.2.0"
regex = "^2020.10.11"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.5.3"


[tool.poetry.dev-dependencies]
black= "^19.10b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
